+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Featherlight – Changelog
 
 Master
 -----------------------------------
-
+- Tweak auto binding to work with persist and remove limitation on filter.
 
 1.3.3 - 2015-08-17
 -----------------------------------